backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
qiskit==1.3.0
# qiskit-aer==0.15.1  # Requires C++ compiler on Windows - install separately if needed
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.0
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.0
google-generativeai==0.3.2
openai==2.32.0
